fix: filter clear button shown without value
Currently the clear button is shown even when the filter value is empty. This commit fixes that by only showing the clear button when the filter value is not empty.
This makes sense because the clear button is meant to clear the filter value, so it should only be shown when there is a value to clear.

Fixes issue: https://github.com/KevinVandy/material-react-table/issues/1284