fix: fixes focus moving to next cell when using arrow keys on slider (#1338)
Previously when using the Arrow Keys for changing the range slider the focus would move to the next/previous cell.
This is now prevented. The only downside is that the user has to press Tab again to move the focus outside of the slider, which in my point should be okay as he has to tab in to get focus. Additionally this is now consistent with text input filters.

Fixes https://github.com/KevinVandy/material-react-table/issues/1254